chat-web-api/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
distro==1.9.0
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.49
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.8
